# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KG-MNP Eligibility API",
    "description": "Frontend-ready backend for deterministic MNP eligibility assessment.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/examples": {
      "get": {
        "tags": [
          "examples"
        ],
        "summary": "List Examples",
        "operationId": "list_examples_api_v1_examples_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/examples/{case_id}": {
      "get": {
        "tags": [
          "examples"
        ],
        "summary": "Get Example",
        "operationId": "get_example_api_v1_examples__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleDetailResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/examples/{case_id}/run": {
      "post": {
        "tags": [
          "examples"
        ],
        "summary": "Run Example",
        "operationId": "run_example_api_v1_examples__case_id__run_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "execution_id": {
            "type": "string",
            "title": "Execution Id"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "assessment_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Time"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EligibilityDecisionCode"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "publication": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicationResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Publication"
          },
          "validations": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationsResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Validations"
          },
          "input_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Summary"
          },
          "evidence": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EvidenceResponse"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Evidence"
          },
          "rule_results": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RuleResultResponse"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Rule Results"
          },
          "blocking_reasons": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BlockingReasonResponse"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Blocking Reasons"
          },
          "remediation_actions": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RemediationActionResponse"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Remediation Actions"
          },
          "process": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProcessResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Process"
          },
          "trace_subgraph": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TraceGraphResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Trace Subgraph"
          },
          "inference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InferenceResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Inference"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "artifacts": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Artifacts"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "schema_version",
          "execution_id",
          "publication",
          "validations"
        ],
        "title": "AssessmentResponse"
      },
...
      "BlockingReasonResponse": {
        "properties": {
          "reason_code": {
            "type": "string",
            "title": "Reason Code"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "rule_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Version"
          },
          "regulatory_clause": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regulatory Clause"
          },
          "action_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Code"
          },
          "evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          },
          "assessment_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Time"
          },
          "effective_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "reason_code"
        ],
        "title": "BlockingReasonResponse"
      },
...
      "EligibilityDecisionCode": {
        "type": "string",
        "enum": [
          "ELIGIBLE",
          "BLOCKED",
          "CONDITIONAL",
          "MANUAL_REVIEW"
        ],
        "title": "EligibilityDecisionCode"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {},
            "type": "array",
            "title": "Details"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "EvidenceResponse": {
        "properties": {
          "evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Id"
          },
          "evidence_iri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Iri"
          },
          "evidence_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Type"
          },
          "source_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source System"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "EvidenceResponse"
      },
...
      "ExampleDetailResponse": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "runnable": {
            "type": "boolean",
            "title": "Runnable",
            "default": true
          },
          "input_format": {
            "type": "string",
            "title": "Input Format",
            "default": "json"
          },
          "expected_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Decision"
          },
          "scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario"
          },
          "input": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input"
          },
          "ttl_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ttl File"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "case_id"
        ],
        "title": "ExampleDetailResponse"
      },
      "ExampleItem": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "runnable": {
            "type": "boolean",
            "title": "Runnable",
            "default": true
          },
          "input_format": {
            "type": "string",
            "title": "Input Format",
            "default": "json"
          },
          "expected_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Decision"
          },
          "scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario"
          },
          "json_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Json File"
          },
          "ttl_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ttl File"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "case_id"
        ],
        "title": "ExampleItem"
      },
      "ExampleListResponse": {
        "properties": {
          "items": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExampleItem"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ExampleListResponse"
      },
...
      "InferenceResponse": {
        "properties": {
          "triples_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triples Before"
          },
          "triples_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triples After"
          },
          "triples_added": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triples Added"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "InferenceResponse"
      },
...
      "ProcessBlockingReason": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ProcessBlockingReason"
      },
...
      "ProcessResponse": {
        "properties": {
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "next_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Step"
          },
          "can_advance": {
            "type": "boolean",
            "title": "Can Advance",
            "default": false
          },
          "blocking_reasons": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProcessBlockingReason"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Blocking Reasons"
          },
          "authorization_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthorizationCodeResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorization Code"
          },
          "eligibility_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eligibility Decision"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ProcessResponse"
      },
      "PublicationResponse": {
        "properties": {
          "publishable": {
            "type": "boolean",
            "title": "Publishable"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicationStatus"
              },
              {
                "type": "string"
              }
            ],
            "title": "Status"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "publishable",
          "status"
        ],
        "title": "PublicationResponse"
      },
      "PublicationStatus": {
        "type": "string",
        "enum": [
          "PUBLISHABLE",
          "NOT_PUBLISHABLE"
        ],
        "title": "PublicationStatus"
      },
...
      "RemediationActionResponse": {
        "properties": {
          "action_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Code"
          },
          "iri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iri"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "RemediationActionResponse"
      },
...
      "RuleResultResponse": {
        "properties": {
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "effective_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "selected_for_assessment_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected For Assessment Time"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "rule_id"
        ],
        "title": "RuleResultResponse"
      },
...
      "TraceGraphResponse": {
        "properties": {
          "nodes": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TraceNodeResponse"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TraceEdgeResponse"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TraceGraphResponse"
      },
...
      "ValidationStepResponse": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "conforms": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conforms"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ValidationStepResponse"
      },
      "ValidationsResponse": {
        "properties": {
          "json_schema": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationStepResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Json Schema"
          },
          "input_graph": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationStepResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Graph"
          },
          "assessment_graph": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationStepResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Graph"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ValidationsResponse"
      },
...
    }
  }
}